fix(core): resolve CI red on review-fixes/core (tach, module-health, docs-drift)
Merge origin/main and fix every failing check without any suppression:

- lint/tach: LeaseLostError moved off the forbidden teatree.agents ->
  teatree.core.models.errors edge. Re-export it from teatree.core.models
  (allowed) and import it there; drop the direct errors import in
  agents/headless.py and models/task.py's caller path.
- module-health-gate:
  - Split the FSM phase-transition disposition/escalation concern out of the
    task.py god-module into models/task_phase_disposition.py (source-state
    derivation, unshippable-review auto-ignore, wedge escalation), bringing
    task.py back under the shrink ratchet.
  - Split the result parse/validate/schema concern out of headless.py into
    agents/headless_result.py, bringing headless.py back under the 500 cap.
  - Replace dict[str, object] with a named JSONObject alias
    (models/types.py) in merge_extra / _record_pr_url.
- docs-drift: regenerate docs/generated/cli-reference.md for the tasks-list
  docstring this branch changed.

Extractions are behaviour-preserving; callers and tests updated to the new
module homes.
