+++ src/teatree/agents/attempt_recorder.py
@@ -110,7 +110,7 @@ def validate_result_keys(result: AgentResultBlob) -> str:
     """Return an error message if *result* carries keys outside the schema.
 
     Only the ``additionalProperties: false`` rule is enforced (no full
-    JSON-Schema dependency), mirroring the headless path's ``_validate_result``.
+    JSON-Schema dependency), mirroring the headless path's ``headless_result.validate_result``.
     """
     allowed = set(RESULT_JSON_SCHEMA.get("properties", {}).keys())  # type: ignore[union-attr]
     unexpected = set(result) - allowed
